Update
Moving towards portability. Segment quantity can now be changed and the
scroll functions should (hopefully) still work well in a 1-dimensional
scroll call, regardless of display segment quantity.

Added some punctuation symbols and other symbol characters to the actual
LCD_Launchpad.cpp, since we desired the use of commas in the n-back
game, and we will also require the use of the negative sign without the
need to be restricted to the "MINUS1" segment of the LCD display.